+++ source/Dgraph-dotnet.examples/MutationExample/Program.cs
@@ -43,7 +43,7 @@ namespace MutationExamples {
 
                     // use Upsert to test for a node and value, and create if
                     // not already in the graph as an atomic operation.
-                    var result = await client.Upsert("Username", GraphValue.BuildStringValue(username));
+                    var result = await client.Upsert("Username", GraphValue.BuildStringValue(username), $"{{\"Username\": \"{username}\"}}");
 
                     if (result.IsFailed) {
                         Console.WriteLine("Something went wrong : " + result);